refactor: map 관련 util 함수 분리 및 리팩토링 (#88)
