+++ src/constants/map.constants.ts
@@ -1,3 +1,5 @@
+import { KakaoMapOptions } from '@/types/kakao-map.type';
+
 // 마커 관련 상수 정의
 export const MARKER = {
   SIZE: 36,
@@ -6,3 +8,9 @@ export const MARKER = {
     Y: 36,
   },
 } as const;
+
+// 초기 설정 map option
+export const DEFAULT_MAP_OPTIONS: KakaoMapOptions = {
+  center: { lat: 33.3616666, lng: 126.5291666 },
+  level: 10,
+};
